feat(echo): make  method public
Due this method needed to be used in the various sections of the project such as extension part of the mechanism, the method under the corresponding name has been changed its visibility from private to public.
dquote> _getUniqueId(dynamic suffix)
dquote> for more information about this method goto the main class of the package (#472).