feat(client): remove features individual registration, instead move into one manage place.
handling roster updates properly was the primary problem in the client, but in this commit it should end up with good results.

add stream features handling delay between requests due processing is so fast and in the case where we need to catch up registration request, we were not able to see it. it should conclude this problem.

provide updateRoster method ot properly update current roster.